Crate ffts

ffts-rs provides Rust bindings for FFTS

Complex data is stored in the interleaved format (i.e, the real and imaginary parts composing each element of complex data are stored adjacently in memory).

The multi-dimensional arrays passed are expected to be stored as a single contiguous block in row-major order.

For real transforms, dir == Dir::Forward implies a real-to-complex forwards tranform, and dir == Dir::Backward implies a complex-to-real backwards transform.

The output of a real-to-complex transform is N/2+1 complex numbers, where the redundant outputs have been omitted.

Structs

Ffts

Enums

Dir
The direction of the transform (i.e, the sign of the exponent in the transform.)